+++ R/master.R
@@ -1,63 +1,76 @@
 # This master file runs all files
-# 20190617 by JJAV
+# 20200513 by JJAV
 # # # # # # # # # # # # # # # # # # # # #
 
-#' Runs the programs
+#' Render programs
 #'
-#' Run the programs specified by the pattern, in the order as the
-#' pattern select the files. Keep a log of the results. By default
-#' the pattern starting with two numbers and ending with .R is selected.
-#' They are run in order
+#' By default, all programs with the pattern "nn_" will be executed in order.
+#' The 'Start' and 'Stop' parameters can be used to modify the files to start or stop at
+#' a different number.
 #'
-#' The program add (or create if not exists) the files run, time of execution
-#' and exit status in the file master.log of the directory logs.
-
-#' @param pattern Regular expression to select the files to run
-#' @param start index of the program to start
-#' @param stop index of the program to stop
+#' The files are treated as _snip_ files for rmarkdown and render in the log
+#' directory, with the format specified in the _format_ parameter.
+#'
+#' The default the log directory is configured in the config.yml file.
+#'
+#' @param start program to start
+#' @param stop  program to stop
+#' @param format Format to render the programs values accepted are "pdf", "html" and "word"
 #' @param logdir directory to keep the logs of the files. By default
-#' @param rscript_path path to the \code{Rscript} file
 #' the entry on the \code{config.yml} \code{dirs:logs}
-#' @import processx
 #' @import readr
 #' @import config
 #' @importFrom utils write.table
-#' @importFrom digest digest
 #' @importFrom tools file_path_sans_ext
+#' @importFrom yaml yaml.load
+#' @importFrom digest digest
 #' @export
 #' @return a data.frame with the files run, running time and exit status
 master <-
-  function(pattern = "^[0-9][0-9].*\\.R$",
-           start = 1,
+  function(
+           start = 0,
            stop = Inf,
-           logdir = config::get("dirs")$logs,
-           rscript_path) {
-    if (missing(rscript_path)) {
-      binpath <- file.path(R.home(), "bin")
-      if (file.exists(file.path(binpath, "Rscript"))) {
-        rscript_path = file.path(binpath, "Rscript")
-      }
-      else {
-        if (file.exists(file.path(binpath, "Rscript.exe"))) {
-          rscript_path = file.path(
-            binpath,
-            "Rscript.exe")
-          }
-        else {
-          rscript_path = "Rscript"
-        }
+           format = "html",
+           logdir = config::get("dirs")$logs) {
 
-      }
-    }
-    line80 <- paste0("'",paste0(rep("=",80), collapse = ""),"\n","'")
+    # The patter is of files starting with two digits
+    pattern = "^[0-9][0-9].*\\.R$"
+
+    # list of files to run
     scriptlist = dir(".", pattern = pattern, full.names = T)
-    tostop = min(length(scriptlist), stop)
+
+    if (! is.infinite(stop)) {
+      numdef <- as.numeric(sub("^\\D+(\\d+).*", "\\1", scriptlist))
+      stopidx <- which(numdef == stop)
+      if (length(stopidx) == 0) {
+        stop("Stop file ", stop, " Not found!")
+      }
+      if (length(stopidx) > 1) {
+        stop("More than one file " , stop, "are defined" )
+      }
+      stop = stopidx
+    }
+    else(
+      stop = length(scriptlist)
+    )
+
+    if (start > 0) {
+      numdef <- as.numeric(sub("^\\D+(\\d+).*", "\\1", scriptlist))
+      startidx <- which(numdef == start)
+      if (length(startidx) == 0) {
+        stop("Start file ", start, " Not found!")
+      }
+      if (length(startidx) > 1) {
+        stop("More than one file " , start, "are defined" )
+      }
+      start = startidx
+    }
+
     reslogs <-
-      lapply(scriptlist[start:tostop], function(x) {
-        cat(x, "\n")
+      lapply(scriptlist[start:stop], function(x) {
+        cat("\n",x, "\n")
+
         # Define if a signature and session infor should be included
-        need_a_signature <- TRUE
-        need_a_sessioninfo <- TRUE
         rlx <- readLines(x)
         headlin <- grep("\\#' ---", rlx)
         # No well format heading
@@ -66,28 +79,19 @@ master <-
           nedd_a_sessioninfo = FALSE
         }
         else {
-          yamlx <- yaml.load(
+          yamalx <- yaml.load(
             gsub("\\#' ", "", rlx[seq(headlin[1] + 1, headlin[2] - 1)]))
-          if (exists("yamalx$signature")) {
-            need_a_signature = identical(yamalx$signature, TRUE)
-          }
-          else {
-            need_a_signature = FALSE
-          }
-
-          if (exists("yamalx$sessioninfo")) {
-            need_a_sessioninfo = identical(yamalx$sessioninfo, TRUE)
-          }
-          else {
-            need_a_sessioninfo = FALSE
-          }
+          need_a_signature = identical(yamalx[["signature"]],TRUE)
+          need_a_sessioninfo = identical(yamalx[["sessioninfo"]],TRUE)
         }
 
 
         # Add a Session info is required
         if (need_a_sessioninfo) {
-          rlx[length(rlx) + 1] <- paste0(line80,
-                                         "# Session Info \n",
+          rlx[length(rlx) + 1] <- paste0("#' ## Session Info ",
+                                         "\n",
+                                         "#+ echo = F ",
+                                         "\n",
                                          "print(sessionInfo(), locale = F)")
         }
 
@@ -101,8 +105,9 @@ master <-
                                         start_var, "<- Sys.time()\n",
                                         rlx[headlin[2] + 1])
           rlx[length(rlx) + 1] <- paste0(
-            line80,
-            "# Signature ",
+            "#' ## Signature ",
+            "\n",
+            "#+ echo = F ",
             "\n",
             "cat(",
             "\n",
@@ -127,53 +132,35 @@ master <-
         fnx <- paste0(basename(tempfile()), ".R")
         bnx <- file_path_sans_ext(basename(x))
         writeLines(rlx, fnx)
+
         start = Sys.time()
-        cat(line80)
         res <-
-          try(processx::run(rscript_path,
-                            c(fnx, "--vainilla"),
-                            echo = T,
-                            error_on_status = FALSE))
+          try(
+            render_report(
+              fnx,
+              format,
+              logdir,
+              quiet = TRUE,
+              output_file = file.path(logdir, bnx)))
+        file.remove(fnx)
+        res <- ifelse(is.null(res), 0, res)
         end <- Sys.time()
         elapsed = difftime(end, start)
-        file.remove(fnx)
-        if (!inherits(res, "try-error")) {
-          fname =  file.path("logs", paste0(x, ".log"))
 
-          readr::write_lines(res$stdout, fname, append = T)
-          cat(line80, file = fname, append = T)
-          if (res$stderr != "") {
-            cat("Errors and Warnings messages:\n\n",
-                file = fname,
-                append = T)
-            readr::write_lines(gsub("\033\\[[0-9]{1,2}m", "", res$stderr),
-                               fname,
-                               append = T)
-          }
-          cat(line80, file = fname, append = T)
-          cat(
-            "Status: ",
-            res$status,
-            "  Timeout: ",
-            res$timeout,
-            "  Time: ",
-            format(elapsed),
-            file = fname,
-            append = T
-          )
-          cat(line80)
-          cat(x, "  ", format(elapsed), "\n")
-        }
+        print(format(start))
+        print(x)
+        print(format(elapsed))
+        print(ifelse(inherits(res, "try-error") | res != 0, "FAIL", ":-)"))
         data.frame(
           timestart = format(start),
           script = x,
           elapsed = format(elapsed),
-          comments = ifelse(inherits(res, "try-error") |
-                              res$status != 0, "FAIL", ":-)")
+          comments = ifelse(inherits(res, "try-error") | res != 0, "FAIL", ":-)")
         )
       })
 
     dfres <- do.call(rbind, reslogs)
+
     # Suppress a warning when append the column names
     suppressWarnings(
       write.table(