+++ tests/testthat/02_prog_a.R
@@ -1,6 +1,10 @@
-# Program test A
-# 20181124 by JJAV
-# # # # # # # # # # #
+#' ---
+#' title: Program Test A
+#' author: JJAV
+#' date: 2023-10-01
+#' sessioninfo: YES
+#' signature: YES
+#' ---
 
 var_a <- "AAA"
 var_b <- "BBB"