+++ tests/testthat/03_prog_b.R
@@ -1,6 +1,10 @@
-# Program test B
-# 20181124 by JJAV
-# # # # # # # # # # #
+#' ---
+#' title: Program Test B
+#' author: JJAV
+#' date: 2023-10-01
+#' sessioninfo: YES
+#' signature: YES
+#' ---
 
 var_a <- "AAA"
 stopifnot(!exists("var_b", inherits = TRUE))