+++ tests/testthat/04_prog_c.R
@@ -1,6 +1,10 @@
-# Program test c
-# 20181124 by JJAV
-# # # # # # # # # # #
+#' ---
+#' title: Program Test C
+#' author: JJAV
+#' date: 2023-10-01
+#' sessioninfo: YES
+#' signature: YES
+#' ---
 
 # This program conclude with an error
 cat("I'm in program C and would fail!!")