+++ tests/testthat/05_prog_d.R
@@ -1,6 +1,10 @@
-# Program test d
-# 20181125 by JJAV
-# # # # # # # # # # #
-library(repana)
+#' ---
+#' title: Program Test D
+#' author: JJAV
+#' date: 2023-10-01
+#' sessioninfo: YES
+#' signature: YES
+#' ---
 
+library(repana)
 cat("I'm in program D")